+++ sentry/src/main/java/io/sentry/MainEventProcessor.java
@@ -1,6 +1,7 @@
 package io.sentry;
 
 import io.sentry.protocol.SentryException;
+import io.sentry.protocol.User;
 import io.sentry.util.ApplyScopeUtils;
 import io.sentry.util.Objects;
 import java.util.ArrayList;
@@ -14,6 +15,12 @@ import org.jetbrains.annotations.Nullable;
 public final class MainEventProcessor implements EventProcessor {
 
   /**
+   * Default value for {@link User#getIpAddress()} set when event does not have user and ip address
+   * set and when {@link SentryOptions#isSendDefaultPii()} is set to true.
+   */
+  public static final String DEFAULT_IP_ADDRESS = "{{auto}}";
+
+  /**
    * Default value for {@link SentryEvent#getEnvironment()} set when both event and {@link
    * SentryOptions} do not have the environment field set.
    */
@@ -22,9 +29,16 @@ public final class MainEventProcessor implements EventProcessor {
   private final @NotNull SentryOptions options;
   private final @NotNull SentryThreadFactory sentryThreadFactory;
   private final @NotNull SentryExceptionFactory sentryExceptionFactory;
+  private final @Nullable HostnameCache hostnameCache;
 
   MainEventProcessor(final @NotNull SentryOptions options) {
+    this(options, options.isAttachServerName() ? new HostnameCache() : null);
+  }
+
+  MainEventProcessor(
+      final @NotNull SentryOptions options, final @Nullable HostnameCache hostnameCache) {
     this.options = Objects.requireNonNull(options, "The SentryOptions is required.");
+    this.hostnameCache = hostnameCache;
 
     final SentryStackTraceFactory sentryStackTraceFactory =
         new SentryStackTraceFactory(
@@ -37,12 +51,14 @@ public final class MainEventProcessor implements EventProcessor {
   MainEventProcessor(
       final @NotNull SentryOptions options,
       final @NotNull SentryThreadFactory sentryThreadFactory,
-      final @NotNull SentryExceptionFactory sentryExceptionFactory) {
+      final @NotNull SentryExceptionFactory sentryExceptionFactory,
+      final @NotNull HostnameCache hostnameCache) {
     this.options = Objects.requireNonNull(options, "The SentryOptions is required.");
     this.sentryThreadFactory =
         Objects.requireNonNull(sentryThreadFactory, "The SentryThreadFactory is required.");
     this.sentryExceptionFactory =
         Objects.requireNonNull(sentryExceptionFactory, "The SentryExceptionFactory is required.");
+    this.hostnameCache = Objects.requireNonNull(hostnameCache, "The HostnameCache is required");
   }
 
   @Override
@@ -122,6 +138,18 @@ public final class MainEventProcessor implements EventProcessor {
         // if there are no exceptions, exceptions have its own stack traces.
         event.setThreads(sentryThreadFactory.getCurrentThread());
       }
+    }
+    if (options.isSendDefaultPii()) {
+      if (event.getUser() == null) {
+        final User user = new User();
+        user.setIpAddress(IpAddressUtils.DEFAULT_IP_ADDRESS);
+        event.setUser(user);
+      } else if (event.getUser().getIpAddress() == null) {
+        event.getUser().setIpAddress(IpAddressUtils.DEFAULT_IP_ADDRESS);
+      }
+    }
+    if (options.isAttachServerName() && hostnameCache != null && event.getServerName() == null) {
+      event.setServerName(hostnameCache.getHostname());
     }
   }
 }