+++ sentry/src/main/java/io/sentry/SentryClient.java
@@ -161,7 +161,7 @@ public final class SentryClient implements ISentryClient {
     if (attachments != null) {
       for (final Attachment attachment : attachments) {
         final SentryEnvelopeItem attachmentItem =
-            SentryEnvelopeItem.fromAttachment(options.getLogger(), attachment);
+            SentryEnvelopeItem.fromAttachment(attachment, options.getMaxAttachmentSize());
         envelopeItems.add(attachmentItem);
       }
     }
@@ -353,7 +353,8 @@ public final class SentryClient implements ISentryClient {
     SentryId sentryId = transaction.getEventId();
 
     if (transaction instanceof SentryTransaction) {
-      final SentryTransaction sentryTransaction = (SentryTransaction) transaction;
+      final SentryTransaction sentryTransaction =
+          processTransaction((SentryTransaction) transaction);
       try {
         final SentryEnvelope envelope =
             buildEnvelope(sentryTransaction, getAttachmentsFromScope(scope));
@@ -374,6 +375,17 @@ public final class SentryClient implements ISentryClient {
     }
 
     return sentryId;
+  }
+
+  private @NotNull SentryTransaction processTransaction(
+      final @NotNull SentryTransaction transaction) {
+    if (transaction.getRelease() == null) {
+      transaction.setRelease(options.getRelease());
+    }
+    if (transaction.getEnvironment() == null) {
+      transaction.setEnvironment(options.getEnvironment());
+    }
+    return transaction;
   }
 
   private @Nullable SentryEvent applyScope(