+++ sentry/src/main/java/io/sentry/exception/SentryEnvelopeException.java
@@ -1,0 +1,10 @@
+package io.sentry.exception;
+
+public final class SentryEnvelopeException extends Exception {
+
+  private static final long serialVersionUID = -8307801916948173232L;
+
+  public SentryEnvelopeException(String message) {
+    super(message);
+  }
+}