+++ sentry/src/test/java/io/sentry/SentryEnvelopeItemTest.kt
@@ -1,15 +1,14 @@
 package io.sentry
 
-import com.nhaarman.mockitokotlin2.any
-import com.nhaarman.mockitokotlin2.eq
 import com.nhaarman.mockitokotlin2.mock
-import com.nhaarman.mockitokotlin2.verify
+import io.sentry.exception.SentryEnvelopeException
 import io.sentry.protocol.User
 import io.sentry.test.injectForField
 import java.io.File
 import kotlin.test.AfterTest
 import kotlin.test.Test
 import kotlin.test.assertEquals
+import kotlin.test.assertFailsWith
 import kotlin.test.assertNotNull
 import kotlin.test.assertNull
 import org.junit.Assert.assertArrayEquals
@@ -18,7 +17,12 @@ class SentryEnvelopeItemTest {
 
     private class Fixture {
         val pathname = "hello.txt"
+        val filename = pathname
         val bytes = "hello".toByteArray()
+        val maxAttachmentSize: Long = 5 * 1024 * 1024
+
+        val bytesAllowed = ByteArray(maxAttachmentSize.toInt()) { 0 }
+        val bytesTooBig = ByteArray((maxAttachmentSize + 1).toInt()) { 0 }
     }
 
     private val fixture = Fixture()
@@ -42,23 +46,22 @@ class SentryEnvelopeItemTest {
 
     @Test
     fun `fromAttachment with bytes`() {
-        val bytes = "hello".toByteArray()
-        val attachment = Attachment(bytes, fixture.pathname)
+        val attachment = Attachment(fixture.bytesAllowed, fixture.filename)
 
-        val item = SentryEnvelopeItem.fromAttachment(mock(), attachment)
+        val item = SentryEnvelopeItem.fromAttachment(attachment, fixture.maxAttachmentSize)
 
-        assertAttachment(attachment, bytes, item)
+        assertAttachment(attachment, fixture.bytesAllowed, item)
     }
 
     @Test
     fun `fromAttachment with file`() {
         val file = File(fixture.pathname)
-        file.writeBytes(fixture.bytes)
+        file.writeBytes(fixture.bytesAllowed)
         val attachment = Attachment(file.path)
 
-        val item = SentryEnvelopeItem.fromAttachment(mock(), attachment)
+        val item = SentryEnvelopeItem.fromAttachment(attachment, fixture.maxAttachmentSize)
 
-        assertAttachment(attachment, fixture.bytes, item)
+        assertAttachment(attachment, fixture.bytesAllowed, item)
     }
 
     @Test
@@ -68,21 +71,21 @@ class SentryEnvelopeItemTest {
         file.writeBytes(twoMB)
         val attachment = Attachment(file.absolutePath)
 
-        val item = SentryEnvelopeItem.fromAttachment(mock(), attachment)
+        val item = SentryEnvelopeItem.fromAttachment(attachment, fixture.maxAttachmentSize)
 
         assertAttachment(attachment, twoMB, item)
     }
 
     @Test
     fun `fromAttachment with non existent file`() {
-        val logger = mock<ILogger>()
         val attachment = Attachment("I don't exist", "file.txt")
 
-        val item = SentryEnvelopeItem.fromAttachment(logger, attachment)
+        val item = SentryEnvelopeItem.fromAttachment(attachment, fixture.maxAttachmentSize)
 
-        assertAttachment(attachment,
-                "Reading the attachment ${attachment.pathname} failed, because the file located at " +
-                        "the path is not a file.", item)
+        assertFailsWith<SentryEnvelopeException>("Reading the attachment ${attachment.pathname} failed, because the file located at " +
+                "the path is not a file.") {
+            item.data
+        }
     }
 
     @Test
@@ -93,13 +96,14 @@ class SentryEnvelopeItemTest {
         // On CI it can happen that we don't have the permission to the file permission to read only
         val changedFileReadPermission = file.setReadable(false)
         if (changedFileReadPermission) {
-            val logger = mock<ILogger>()
             val attachment = Attachment(file.path, "file.txt")
 
-            val item = SentryEnvelopeItem.fromAttachment(logger, attachment)
+            val item = SentryEnvelopeItem.fromAttachment(attachment, fixture.maxAttachmentSize)
 
-            assertAttachment(attachment, "Reading the attachment ${attachment.pathname} failed, " +
-                    "because can't read the file.", item)
+            assertFailsWith<SentryEnvelopeException>("Reading the attachment ${attachment.pathname} failed, " +
+                    "because can't read the file.") {
+                item.data
+            }
         } else {
             println("Was not able to change file access permission. Skipping test.")
         }
@@ -110,16 +114,16 @@ class SentryEnvelopeItemTest {
         val file = File(fixture.pathname)
         file.writeBytes(fixture.bytes)
 
-        val logger = mock<ILogger>()
-        val attachment = Attachment(file.path, "file.txt")
+        val attachment = Attachment(file.path, fixture.filename)
 
         val securityManager = DenyReadFileSecurityManager(fixture.pathname)
         System.setSecurityManager(securityManager)
 
-        val item = SentryEnvelopeItem.fromAttachment(logger, attachment)
+        val item = SentryEnvelopeItem.fromAttachment(attachment, fixture.maxAttachmentSize)
 
-        assertAttachment(attachment, "Reading the attachment ${attachment.pathname} failed.", item)
-        verifyLogException<SecurityException>(logger, attachment.pathname ?: "")
+        assertFailsWith<SentryEnvelopeException>("Reading the attachment ${attachment.pathname} failed.") {
+            item.data
+        }
 
         System.setSecurityManager(null)
     }
@@ -134,10 +138,12 @@ class SentryEnvelopeItemTest {
         // reflection instead.
         attachment.injectForField("pathname", null)
 
-        val item = SentryEnvelopeItem.fromAttachment(mock(), attachment)
+        val item = SentryEnvelopeItem.fromAttachment(attachment, fixture.maxAttachmentSize)
 
-        assertAttachment(attachment, "Couldn't attach the attachment ${attachment.filename}.\n" +
-                "Please check that either bytes or a path is set.", item)
+        assertFailsWith<SentryEnvelopeException>("Couldn't attach the attachment ${attachment.filename}.\n" +
+                "Please check that either bytes or a path is set.") {
+            item.data
+        }
     }
 
     @Test
@@ -145,8 +151,36 @@ class SentryEnvelopeItemTest {
         val image = this::class.java.classLoader.getResource("Tongariro.jpg")!!
         val attachment = Attachment(image.path)
 
-        val item = SentryEnvelopeItem.fromAttachment(mock(), attachment)
+        val item = SentryEnvelopeItem.fromAttachment(attachment, fixture.maxAttachmentSize)
         assertAttachment(attachment, image.readBytes(), item)
+    }
+
+    @Test
+    fun `fromAttachment with bytes too big`() {
+        val attachment = Attachment(fixture.bytesTooBig, fixture.filename)
+        val exception = assertFailsWith<SentryEnvelopeException> {
+            SentryEnvelopeItem.fromAttachment(attachment, fixture.maxAttachmentSize).data
+        }
+
+        assertEquals("Dropping attachment with filename '${fixture.filename}', because the " +
+                "size of the passed bytes with ${fixture.bytesTooBig.size} bytes is bigger " +
+                "than the maximum allowed attachment size of " +
+                "${fixture.maxAttachmentSize} bytes.", exception.message)
+    }
+
+    @Test
+    fun `fromAttachment with file too big`() {
+        val file = File(fixture.pathname)
+        file.writeBytes(fixture.bytesTooBig)
+        val attachment = Attachment(file.path)
+
+        val exception = assertFailsWith<SentryEnvelopeException> {
+            SentryEnvelopeItem.fromAttachment(attachment, fixture.maxAttachmentSize).data
+        }
+
+        assertEquals("Dropping attachment, because the size of the it located at " +
+                "'${fixture.pathname}' with ${file.length()} bytes is bigger than the maximum " +
+                "allowed attachment size of ${fixture.maxAttachmentSize} bytes.", exception.message)
     }
 
     private fun createSession(): Session {
@@ -161,21 +195,5 @@ class SentryEnvelopeItemTest {
         assertEquals(attachment.contentType, actualItem.header.contentType)
         assertEquals(attachment.filename, actualItem.header.fileName)
         assertArrayEquals(expectedBytes, actualItem.data)
-    }
-
-    private fun assertAttachment(
-        attachment: Attachment,
-        expectedErrorMessage: String,
-        actualItem: SentryEnvelopeItem
-    ) {
-        assertEquals(attachment.contentType, actualItem.header.contentType)
-        assertEquals(attachment.filename, actualItem.header.fileName)
-        assertEquals(expectedErrorMessage, String(actualItem.data))
-    }
-
-    private inline fun <reified T : Exception> verifyLogException(logger: ILogger, pathname: String) {
-        verify(logger)
-            .log(eq(SentryLevel.ERROR), any<T>(),
-                eq("Reading the attachment $pathname failed."))
     }
 }