+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # vNext
 
+* Feat: Add maxAttachmentSize to SentryOptions (#1138)
+* Feat: Drop invalid attachments (#1134)
 * Ref: Make Attachment immutable (#1120)
 * Fix inheriting sampling decision from parent (#1100)
 * Fixes and Tests: Session serialization and deserialization
@@ -14,7 +16,14 @@
 * Ref: Add option to set `TransportFactory` instead of `ITransport` on `SentryOptions` (#1124)
 * Ref: Simplify ITransport creation in ITransportFactory (#1135) 
 * Feat: Add non blocking Apache HttpClient 5 based Transport (#1136)
-* Enhancement: Autoconfigure Apache HttpClient 5 based Transport in Spring Boot integration (#1143)   
+* Enhancement: Autoconfigure Apache HttpClient 5 based Transport in Spring Boot integration (#1143)
+* Enhancement: Send user.ip_address = {{auto}} when sendDefaultPii is true (#1015)
+* Fix: Resolving dashed properties from external configuration
+* Feat: Read `uncaught.handler.enabled` property from the external configuration 
+* Feat: Resolve servername from the localhost address
+* Fix: Consider {{ auto }} as a default ip address (#1015) 
+* Fix: Set release and environment on Transactions (#1152)
+* Fix: Do not set transaction on the scope automatically   
 
 # 4.0.0-alpha.2